renamed _sentiment to _user and updated refs to it
