+++ tests.py
@@ -1,5 +1,5 @@
 import hacker_access
-import hacker_sentiment
+import hacker_user
 import sentiment_model
 import main
 import heroku_pass_off
